pyproject.toml
[project]
name = "toeic4all-admin"
version = "0.1.0"
description = "Toeic Question Generation by LLM"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.15.0",
    "pydantic-settings>=2.9.1",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = []
